[#39] replace Falcon by TalonFX
- [x] Head over to the falcon directory in motorcontrol rename FalconVelocityController to TalonFXVelocity controller and FalconPositionController to TalonFXPositionController.
- [x] Rename the directory to Talon
- [x] Replace falon by TalonFX in relevant files
