+++ .gitignore
@@ -80,6 +80,7 @@ hs_err_pid*
 .DS_Store
 .AppleDouble
 .LSOverride
+.history
 
 # Icon must end with two \r
 Icon
